Moved local server to separate class
