+++ lib/logger.js
@@ -55,7 +55,6 @@ CordovaLogger.ERROR = 'error';
 CordovaLogger.RESULTS = 'results';
 
 CordovaLogger.prototype.log = function (logLevel, message) {
-    debugger;
     // if there is no such logLevel defined, or provided level has
     // less severity than active level, then just ignore this call and return
     if (!this.levels[logLevel] || this.levels[logLevel] < this.levels[this.logLevel])
