+++ lib/paramedic.js
@@ -1,11 +1,11 @@
 #!/usr/bin/env node
 
-var http = require('http'),
-    localtunnel = require('localtunnel'),
+var localtunnel = require('localtunnel'),
     parseArgs = require('minimist'),
     shell = require('shelljs'),
     fs = require('fs'),
     request = require('request'),
+    Server = require('./LocalServer'),
     Q = require('q'),
     tmp = require('tmp'),
     PluginsManager = require('./PluginsManager'),
@@ -68,8 +68,8 @@ ParamedicRunner.prototype = {
             return;
         }
 
-        this.startServer()
-        .then(function (server) {
+        logger.info("cordova-paramedic: starting local medic server");
+        new Server(this).startServer().then(function (server) {
 
             self.targets.reduce(function (promise, target) {
 
@@ -152,61 +152,6 @@ ParamedicRunner.prototype = {
         }
         else {
             logger.error("Oops, could not find config.xml");
-        }
-    },
-    startServer: function() {
-        var runner = this;
-        return Q.promise(function (resolve, reject) {
-            logger.info("cordova-paramedic: starting local medic server " + this.platformId);
-            http.createServer(runner.requestListener.bind(runner))
-            .listen(runner.port, '127.0.0.1',function (server) {
-                resolve(server);
-            });
-        });
-    },
-    requestListener: function(request, response) {
-        var self = this;
-        if (request.method == 'PUT' || request.method == 'POST') {
-            var body = '';
-            request.on('data', function (data) {
-                body += data;
-                // Too much POST data, kill the connection!
-                if (body.length > 1e6) {
-                    req.connection.destroy();
-                }
-            });
-            request.on('end', function (res) {
-                if(body.indexOf("mobilespec")  == 2){ // {\"mobilespec\":{...}}
-                    try {
-                        var results = JSON.parse(body);
-                        logger.info("Results: ran " +
-                                        results.mobilespec.specs +
-                                        " specs with " +
-                                        results.mobilespec.failures +
-                                        " failures");
-                        if(results.mobilespec.failures > 0) {
-                            self.cleanUpAndExitWithCode(1,results);
-                        }
-                        else {
-                            self.cleanUpAndExitWithCode(0,results);
-                        }
-
-                    }
-                    catch(err) {
-                        logger.error("parse error :: " + err);
-                        self.cleanUpAndExitWithCode(1);
-                    }
-                }
-                else {
-                    logger.normal("console-log:" + body);
-                }
-            });
-        }
-        else {
-            logger.verbose(request.method);
-            response.writeHead(200, { 'Content-Type': 'text/plain'});
-            response.write("Hello"); // sanity check to make sure server is running
-            response.end();
         }
     },
     addAndRunPlatform: function(target) {
